Modifications requested

+++ db/migrate/20180126080839_create_estudiantes.rb
@@ -4,10 +4,10 @@ class CreateEstudiantes < ActiveRecord::Migration[5.1]
       t.string :nombre
       t.string :ncuenta
       t.string :correo_electronico
-      t.decimal :exact_sigma
-      t.decimal :exact_mean
-      t.decimal :real_sigma
-      t.decimal :real_mean
+      t.float :exact_sigma
+      t.float :exact_mean
+      t.float :real_sigma
+      t.float :real_mean
       t.decimal :seed
       t.text :data
       t.timestamps
